fix(package): avoid action affordance import cycle
Problem: adding ActionAffordancePayload directly from operations into surfaces.payloads made polylogue-mcp startup cycle through operations.__init__ before SurfacePayloadModel was initialized, which failed distribution and Nix package smoke checks.

What changed: move the pure action-affordance Pydantic DTOs into polylogue.surfaces.action_affordances, re-export them through operations.action_contracts for existing callers, and refresh the topology projection for the new surface module.

Verification: uv run python -c 'import polylogue.mcp.cli; import polylogue.surfaces.payloads'; devtools test tests/unit/surfaces/test_search_envelope_contract.py tests/unit/cli/test_cli_action_contracts.py::test_action_contracts_emit_shared_affordance_payloads tests/unit/cli/test_cli_action_contracts.py::test_query_result_actions_declare_selection_or_confirmation_affordance tests/unit/mcp/test_tool_contracts.py::TestActionAffordancesTool::test_action_affordances_returns_shared_payload -q; devtools render all; devtools verify --quick
