+++ docs/plans/topology-target.yaml
@@ -289,7 +289,7 @@ files:
     owner: archive-query
     reason: archive-domain query semantics
   - path: polylogue/archive/query/completions.py
-    loc: 517
+    loc: 520
     target: polylogue/archive/query/completions.py
     owner: archive-query
     reason: archive-domain query semantics
@@ -717,11 +717,11 @@ files:
     target: polylogue/cli/commands/check.py
     owner: stable
   - path: polylogue/cli/commands/completions.py
-    loc: 85
+    loc: 93
     target: polylogue/cli/commands/completions.py
     owner: stable
   - path: polylogue/cli/commands/config.py
-    loc: 121
+    loc: 126
     target: polylogue/cli/commands/config.py
     owner: stable
   - path: polylogue/cli/commands/dashboard.py
@@ -1243,7 +1243,7 @@ files:
     target: polylogue/daemon/healthz.py
     owner: stable
   - path: polylogue/daemon/http.py
-    loc: 3537
+    loc: 3547
     target: polylogue/daemon/http.py
     owner: stable
   - path: polylogue/daemon/live_ingest_attempt_models.py
@@ -1304,7 +1304,7 @@ files:
     target: polylogue/daemon/provenance.py
     owner: stable
   - path: polylogue/daemon/route_contracts.py
-    loc: 461
+    loc: 470
     target: polylogue/daemon/route_contracts.py
     owner: stable
   - path: polylogue/daemon/similarity.py
@@ -1631,7 +1631,7 @@ files:
     target: polylogue/mcp/server_support.py
     owner: stable
   - path: polylogue/mcp/server_tools.py
-    loc: 1040
+    loc: 1054
     target: polylogue/mcp/server_tools.py
     owner: stable
   - path: polylogue/operations/__init__.py
@@ -1639,7 +1639,7 @@ files:
     target: polylogue/operations/__init__.py
     owner: stable
   - path: polylogue/operations/action_contracts.py
-    loc: 225
+    loc: 350
     target: polylogue/operations/action_contracts.py
     owner: stable
   - path: polylogue/operations/archive.py
@@ -3419,8 +3419,12 @@ files:
     loc: 0
     target: polylogue/surfaces/__init__.py
     owner: stable
+  - path: polylogue/surfaces/action_affordances.py
+    loc: 71
+    target: polylogue/surfaces/action_affordances.py
+    owner: stable
   - path: polylogue/surfaces/payloads.py
-    loc: 2527
+    loc: 2538
     target: polylogue/surfaces/payloads.py
     owner: stable
   - path: polylogue/telemetry/__init__.py
