+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 684
+- **Stable** (no move scoped): 685
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 825
-- **Realized polylogue/**/*.py**: 825 files declared
+- **Total declared**: 826
+- **Realized polylogue/**/*.py**: 826 files declared
 <!-- END GENERATED: topology-status -->
